git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/BAM-COW@26053 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/cpa/smg/SMGConcreteErrorPathAllocator.java
@@ -344,7 +344,7 @@ public class SMGConcreteErrorPathAllocator {
     private Address nextAlloc = Address.valueOf(BigInteger.valueOf(100));
     private Map<LeftHandSide, Address> variableAddressMap = new HashMap<>();
 
-    public Address calculateAddress(SMGObject pObject, int pOffset,
+    public Address calculateAddress(SMGObject pObject, long pOffset,
         SMGState pSMGState) {
 
       // Create a new base address for the object if necessary
