+++ src/org/sosy_lab/cpachecker/cpa/smg/graphs/object/generic/MaterlisationStep.java
@@ -195,7 +195,7 @@ public class MaterlisationStep {
     /*Fourth, create all values that are contained in this abstraction*/
     for (SMGEdgeHasValueTemplateWithConcreteValue aField : fieldTemplateContainingValue) {
       SMGObjectTemplate templateObject = aField.getObjectTemplate();
-      int offset = aField.getOffset();
+      long offset = aField.getOffset();
       int value = aField.getValue();
       SMGObject concreteObject = concreteObjectMap.get(templateObject);
       CType type = aField.getType();
@@ -250,7 +250,7 @@ public class MaterlisationStep {
       Set<SMGEdgeHasValue> concreteHves) {
     SMGObjectTemplate templateObject = pAbstractField.getObjectTemplate();
     SMGObject object = pConcreteObjectMap.get(templateObject);
-    int offset = pAbstractField.getOffset();
+    long offset = pAbstractField.getOffset();
     int abstractValue = pAbstractField.getAbstractValue();
     int value = pAbstractObjectToPointersMap.get(templateObject).get(abstractValue);
     CType type = pAbstractField.getType();
@@ -263,7 +263,7 @@ public class MaterlisationStep {
       Set<SMGEdgePointsTo> concretePointer) {
 
     SMGObjectTemplate templateTarget = pAbstractPointer.getObjectTemplate();
-    int offset = pAbstractPointer.getOffset();
+    long offset = pAbstractPointer.getOffset();
     int abstractPointerValue = pAbstractPointer.getAbstractValue();
 
     int concretePointerValue =
